feat(publish)!: include map publishing in "publish" command
BREAKING CHANGE: the map command has been removed in favour of a single publish command
